Display correct value for unlimited ulimit
When doing a container inspect on a container with unlimited ulimits,
the value should be -1.  But because the OCI spec requires the ulimit
value to be uint64, we were displaying the inspect values as a uint64 as
well.  Simple change to display as an int64.

Fixes: #9303

+++ libpod/container_inspect.go
@@ -864,8 +864,8 @@ func (c *Container) generateInspectContainerHostConfig(ctrSpec *spec.Spec, named
 		for _, limit := range ctrSpec.Process.Rlimits {
 			newLimit := define.InspectUlimit{}
 			newLimit.Name = limit.Type
-			newLimit.Soft = limit.Soft
-			newLimit.Hard = limit.Hard
+			newLimit.Soft = int64(limit.Soft)
+			newLimit.Hard = int64(limit.Hard)
 			hostConfig.Ulimits = append(hostConfig.Ulimits, newLimit)
 		}
 	}
